start of owtch

+++ include/sprite_util.h
@@ -14,7 +14,7 @@ void SpriteUtilTakeKnockback(u8 spriteSlot);
 u32 SpriteUtilCheckObjectsTouching(u16 o1Top, u16 o1Bottom, u16 o1Left, u16 o1Right, u16 o2Top, u16 o2Bottom, u16 o2Left, u16 o2Right);
 void SpriteUtilSamusAndSpriteCollision(void);
 u32 SpriteUtilCheckVerticalCollisionAtPosition(u16 yPosition, u16 xPosition);
-u16 SpriteUtilCheckVerticalCollisionAtPositionSlopes(u16 yPosition, u16 xPosition);
+u32 SpriteUtilCheckVerticalCollisionAtPositionSlopes(u16 yPosition, u16 xPosition);
 void unk_1129c(void);
 void unk_11310(void);
 void SpriteUtilCheckCollisionAtPosition(u16 yPosition, u16 xPosition);
